added Rototranslation class in utils
