--- scripts/rototranslation/BELV_UTM2LOC.csv
@@ -1,4 +1,0 @@
-0.7073398398761516 0.7076345860745783 -5.388432423248565e-05 -3897208.232165018
--0.7076345676976755 0.7073398458069671 -0.00023220905364489653 -3306208.087588875
--0.00012613677717273136 0.0002022722927644148 1.0005379856449832 -2745.702093307824
-0.0 0.0 0.0 1.0
