+++ scripts/rototranslation/BELV_UTM2LOC.txt
@@ -1,0 +1,4 @@
+0.7073398398761 0.7076345860743 -5.388432423248565e-05 -3897208.2321650
+-0.707634567697 0.7073398458069 -0.000232209053644 -3306208.087588
+-0.000126136777 0.0002022722927 1.0005379856449832 -2745.70209330
+0.0 0.0 0.0 1.0
