+++ scripts/update_dem.py
